requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
curl_cffi==0.13.0
et_xmlfile==2.0.0
frozendict==2.4.7
idna==3.11
multitasking==0.0.12
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
peewee==4.0.1
pillow==12.1.1
platformdirs==4.9.4
protobuf==7.34.0
pycparser==3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
ttkwidgets==0.13.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
